+++ Tests/EntityFrameworkCore.SplitQueries.PostgreSQL/EntityFrameworkCore.SplitQueries.PostgreSQL.Infrastructure/Repositories/GeometryTypeRepository.cs
@@ -1,0 +1,43 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using EntityFrameworkCore.SplitQueries.PostgreSQL.Domain.Entities;
+using EntityFrameworkCore.SplitQueries.PostgreSQL.Domain.Repositories;
+using EntityFrameworkCore.SplitQueries.PostgreSQL.Infrastructure.Persistence;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.EntityFrameworkCore.Repositories.Repository", Version = "1.0")]
+
+namespace EntityFrameworkCore.SplitQueries.PostgreSQL.Infrastructure.Repositories
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class GeometryTypeRepository : RepositoryBase<GeometryType, GeometryType, ApplicationDbContext>, IGeometryTypeRepository
+    {
+        public GeometryTypeRepository(ApplicationDbContext dbContext) : base(dbContext)
+        {
+        }
+
+        public async Task<GeometryType?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default)
+        {
+            return await FindAsync(x => x.Id == id, cancellationToken);
+        }
+
+        public async Task<GeometryType?> FindByIdAsync(
+            Guid id,
+            Func<IQueryable<GeometryType>, IQueryable<GeometryType>> queryOptions,
+            CancellationToken cancellationToken = default)
+        {
+            return await FindAsync(x => x.Id == id, queryOptions, cancellationToken);
+        }
+
+        public async Task<List<GeometryType>> FindByIdsAsync(Guid[] ids, CancellationToken cancellationToken = default)
+        {
+            // Force materialization - Some combinations of .net9 runtime and EF runtime crash with "Convert ReadOnlySpan to List since expression trees can't handle ref struct"
+            var idList = ids.ToList();
+            return await FindAllAsync(x => idList.Contains(x.Id), cancellationToken);
+        }
+    }
+}